+++ src/UI/TransactionsToastList/components/CustomToast/components/index.tsx
@@ -1,2 +1,3 @@
 export * from './SimpleToast';
 export * from './IconToast';
+export * from './CustomComponentToast';